move locking commands to archiver.locks
